device_utils: Update permission denylist with QUERY_ALL_PACKAGES
Trying to set the QUERY_ALL_PACKAGES permission on one of our bots
running Android O results in "Unknown permission:
android.permission.QUERY_ALL_PACKAGES". More details on the bot and the
links to the scenario can be found in the referenced bug.

The permission was added to the manifest in http://crbug.com/1042023.
The use-case for the permission described there does not seem to be hit
by any benchmark.

Alternatively we could set GN arg:
  android_sdk_release = "p",
which would have avoided the permission in the manifest. Here "p" stands
for the last release that Monochrome.apk is shipping to. This method is
less preferable because devs often build Monochrome.apk without this
flag. Also it seems that official bots set the default SDK release even
when building Monochrome, so an unknown permission should be safely
ignored by older platforms.

Bug: chromium:1117184
Change-Id: I9037fb111432c9908aa7221aef724c77d5d389d6
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2398822
Auto-Submit: Egor Pasko <pasko@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>