Allow forcing a clean shutdown via an environment variable.
This CL allows forcing a clean shutdown of the Chrome browser process
when running a benchmark. It's enabled when the
ENSURE_CLEAN_CHROME_SHUTDOWN environment variable is set. It works by
sending a Browser.close command via the browser Devtools connection and
then wait for the Browser to exit.

Meant to be used for PGO profiling builds.

Bug: chromium:1124774
Change-Id: Ib7a7a29ff51a43663b2c5feae213ffc7f064204a
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2411983
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>