pinpoint: Signal a failure rather than No-Repro when no results.
This improves the bug message in the case where a job could not
reproduce a regression because all attempts for a change failed.
Previously the bug message would misleadingly just say “Couldn't
reproduce a difference”, now it will report that the job failed, and
prompt a human to investigate further.

Also, in the debug log: log the actual comparison result, not just its
p-value.

Bug: chromium:1131189
Change-Id: I74735671c540c6b7dac918e8bda6a4f38894c974
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2449790
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Reviewed-by: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>