[typ] Add ResultSink support
Adds support to typ for reporting test results through ResultSink
natively instead of relying on the automatic JSON conversion that
happens on builders.

This does not require any additional flags to be used, as it will
automatically be used in addition to the existing JSON results if
run on swarming with ResultSink running.

Bug: chromium:1104252
Change-Id: Ief26354908fdb8affaef9d358706a75ec06d93cb
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2468776
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>