Change :generate_about_tracing to be a python2_action.
This CL marks //tracing:generate_about_tracing as needing
to be run with Python 2 rather than Python 3. This will
help with migrating the Chromium build from Python 2 to 3.

Ultimately, of course, we need to fix this so that
the script just works with Python 3 regardless.

Bug: chromium:1112471
Change-Id: I2e267cf454a1f2054ddf9b4dea7f1a5f3dcd818f
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2511042
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>