[typ] ResultSink improvements
Makes the following improvements to typ's ResultSink integration:
1. Includes the typ expectation for the test as a ResultDB tag.
2. Manually closes the connection to ResultSink after sending data, as
   apparently it does not happen automatically despite not allowing the
   user to send additional data.

As a side effect, also changes how tags are handled internally before
being reported to ResultDB to be a list of tuples instead of a dict.
This allows us to report multiple values for the same key, which
ResultDB supports.

Bug: chromium:1104252
Change-Id: I3a201dcd8768bbc5373fe8787630e84fc5c397fd
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2515273
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>