[benchmarking] Update signature of CustomizeOptions method of Benchmark class
We will remove the FindBrowser call in CustomizeOptions, in order to avoid the FindBrowser to be called twice during RunBenchmark.

This CL updates the signature of the CustomizaOptions of the parent class in Telemetry.

We will have another CL to update the child class in Perf repo, and then a last CL to update the caller on Telemetry. More details on:
crbug/1149275#c6

Bug: chromium:1149275
Change-Id: I8d6f20c72548635b22b18c11a5f0dcf67ef27e4c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2550943
Auto-Submit: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>