[trace-viewer] Speed up event histogram computation
We were calling document.createElement (an expensive function) twice,
once to check if a tag name is registered, and once to actually create
it for use. Now we call it only once. Local testing on a trace with lot
of events showed a 33% speed improvement.

Upstream of cl/349443600. b/176299939 for context.

Bug: chromium:1036492
Change-Id: I74f24d5cd712eddd4561dd7bb242f58ae6d3c29e
R=khokhlov@google.com
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2611314
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>
Commit-Queue: Deep Roy <dproy@chromium.org>