Set download path to a temp directory
Previously the download path is set to the default location,
which may gradually fill up disk space when running stories with a
download link. This CL addresses the issue by setting to a temp directory
and remove the directory after the story is complete.

Change-Id: I86285cfe679ac8297cb2b845df4feeb6db74dc2e
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2733589
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Yuheng Huang <yuhengh@chromium.org>