chromeperf: Noise calculation dataflow pipeline.
Generates noise_by_* tables BigQuery with kurtosis, skewness, etc values
for measurements in chromeperf over the specified number of days.

Briefly, the pipeline:

1. Scrapes the last N days of sample_values from the 'rows'
   table in BigQuery.
2. Groups all the samples for a measurement together.
3. Summarizes those sample sets with scipy.stats.describe.
4. Writes the result back to BigQuery in noise_by_* tables.

This is a pretty simple process, so a lot of the code here is
boilerplate for massaging data structures into the right shape for each
step.

Change-Id: Ic2588ad663468cb02b976c9cba46e67cff5bd9be
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2838565
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Reviewed-by: Dean Berris <dberris@chromium.org>