[NetLog Viewer] Fix JSZip reference
The prior change added a reference to the JSZip library, but due to the
lack of a leading slash, the vulcanizer could not correctly find and
embed the library. Similarly, the test server script serve_static was
returning a 404 for the file. This CL fixes both issues.

Bug: chromium:1057782
Change-Id: Ie1f2ceea909ba061013449b980d1cbeb45ac063b
NOPRESUBMIT=true
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2969786
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Eric Lawrence [MSFT] <ericlaw@microsoft.com>