[chromeperf] Add auto_triage_enable logging to alert_group_workflow.py
This should have no effect on the behvior of the application other than
logging more information.

I've not found anything in Datastore for these problem alerts that is
consistent with the erroneous issue filing we've seen. I've also been
unable to reproduce the behavior locally in unit tests. Hence, adding
extra logging to try and catch it in the act in production.

Bug: chromium:1223401
Change-Id: I1f46460982ba6f6702129458abb7e455968efe65
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3062099
Reviewed-by: Ryan Heise <heiserya@google.com>
Commit-Queue: Sean McCullough <seanmccullough@google.com>