[Telemetry] Add full logging artifact wrapper
Adds the FullLoggingArtifactCompatibilityWrapper, which is the same as
LoggingArtifactCompatibilityWrapper but logs the entire artifact instead
of only the first 100 characters. This is so that tests can opt in to
the extra information at the cost of increased log spam if the trade-off
is worth it.

Bug: chromium:1249302
Change-Id: I3ce401dbc295ab631544103db7bc6149abeed3c0
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3159362
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>