Add retry in testScreenshot on android
There is a problem on android devices. In 'Navigate' call there is a
check, that page is fully loaded, but right after this check page is
not fully rendered. Screenshot is taken right after 'Navigate' call,
which causes race condition: scrennshot is taken before colored square
is rendered on page. This CL tries to find pixel with required color
several times, before it treats test as failed.

Change-Id: Ib2e0b94826099a1f05d6d8844035f601663fce98
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3281937
Commit-Queue: Alexander Yashkin <a-v-y@yandex-team.ru>
Reviewed-by: John Chen <johnchen@chromium.org>