[fuchsia] Removes user-agent type for web-engine-shell browser
The user-agent flag is unused in WebEngine and would need to be escaped
by quotation marks, based on how the browser is launched.

This would allow benchmarks that require a shared_page_state to be run
(like rendering.mobile).

This is blocking the following CL:
https://chromium-review.googlesource.com/c/chromium/src/+/3289993

Bug: chromium:1269978
Change-Id: I3e4470e1e23c4b44f304e51e4364ae9f96e8c7e4
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3283426
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Rohan Pavone <rohpavone@chromium.org>