Change histogram_proto to use a custom logger.
logging.warn() written in a import statement causes the default logging
in the program to get initialized. Once the logging is initialized there
is no way to change the logging level. The CL updates this to use a
custom logger to the files importing this module can still change the
default logging level.
The log message or the usage of the module does not change.

Bug: chromium:1239750
Change-Id: I55a4628ea5c1eea72142ecbe43c15ed0dce314c7
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3296627
Reviewed-by: Deep Roy <dproy@chromium.org>
Commit-Queue: ssid <ssid@chromium.org>