Clear app data if you would have uninstalled
If an app would have been uninstalled and then reinstalled but instead
it was skipped because it is up to date (same as on the host). This
change clears the app data since that is what would have happened if the
app was uninstalled.

Change-Id: I610b35fde655a2484a322567c0ae904179fd93f5
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3317996
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>