[code-health] Prepare dashboard/dashboard/api for pylint update
This prepares the dashboard/dashboard/api directory for an update to
the 2.7 version of pylint. In addition, this fixes a number of issues
to make the codebase more Python 3-friendly.
This also gives explicit names to additional arguments in
ApiRequestHandler.Post() and properly deletes them on unused
overrides.

Bug: chromium:1262292
Change-Id: Id587edd1ee4585894aef702fe0bee2077dba82cb
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3331204
Reviewed-by: Jeremy Leconte <jleconte@google.com>
Commit-Queue: Fabrice de Gans <fdegans@chromium.org>