android: startup tracing: disable safebrowsing
We recently found that slow responses are a common problem (more details
on the bug).

The crrev.com/i/4561411 introduces this Android-specific command line flag:
  --danger-disable-safebrowsing-for-benchmarking

When used, it simulates a 'realistic' delay from safebrowsing, isolating
very slow responses.

Add this flag in places related to startup benchmarking to (a) see
whether the issue affects the bots in the performance lab, (b) allow us
to focus on orthogonal optimizations while the SB issue is being worked
on.

Bug: chromium:1296097
Change-Id: Ic730204b28c8fca16009891fbea3b00980da72a3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3484062
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>