Catches DevtoolsTargetClosedException when navigating back
Historically content has sent ipcs to the existing renderer
before navigating. This was done to service beforeunload handlers
and was sent even if the page did not have a beforeunload handler.
I'm in the process of making it so content only sends the ipc if
the page has a beforeunload handler. This has exposed a timing
issue with the ActionRunner.

When navigating back the ActionRunner uses history.window.back,
which devtools waits for a response from. The problem is that
history.window.back may immediately navigate, resulting in closing
the old renderer. Devtools detects the close and sends a target-closed
message.

This changes ActionRunner so that when navigating back a
DevtoolsTargetClosedException is caught and assumed to equate to
success.

Bug: chromium:365039,chromium:1277988
Change-Id: I0aefc3ab3c81d58b4a9ef9eca5cef3e7cf915310
TEST=none
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3572230
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>