[typ] Limit RDB test name tag length
Limits the length of the test_name tag given to ResultDB to 256
characters, as this is the limit enforced by ResultDB itself. If the
name is longer than that, the first 216 characters are kept and the SHA1
hash of the name appended to the end.

Bug: chromium:1313564
Change-Id: I870f5db2e79251a5e8f993f497ebb65eefab957c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3578580
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Commit-Queue: Rakib Hasan <rmhasan@google.com>