[pinpoint] migrate handlers to flask: task_updates
This handler should be used for pubsub only for execution engine. However, it is enabled in production and taking traffics.

https://pantheon.corp.google.com/logs/query;query=resource.type%3D%22gae_app%22%20OR%20resource.type%3D%22global%22%20logName%3D%22projects%2Fchromeperf%2Flogs%2Fappengine.googleapis.com%252Frequest_log%22%20resource.labels.module_id%3D%22pinpoint%22%20resource.labels.version_id%3D%22cloud-build-62300d7%22%20severity%3E%3DDEFAULT%20%2528textPayload:%22LOGPOINT:%22%2529%20OR%20%2528protoPayload.line.sourceLocation.file:%22task_updates.py%22%20AND%20protoPayload.line.sourceLocation.line:%2226%22%20AND%20protoPayload.line.logMessage:%22LOGPOINT:%22%2529%0A%22task_updates%22;timeRange=PT15M;cursorTimestamp=2022-05-04T22:08:39.855861Z?project=chromeperf

I updated the handler but have no easy way to test it as no pubsub is set up in staging. I think we can wait to see how it works in prod as there should be nothing to break.

Bug: chromium:1298177
Change-Id: Id5bc075933fd16e1c919cad2619c43c4926fe942
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3627476
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>