Determine target platforms from browser
If not already set, attempts to set
BrowserFinderOptions.target_platforms based on the the value of
--browser. This ends up doing some preliminary pruning of browser
finders in an effort to avoid weird dependency issues.

For example, running a test with --browser=release (a desktop browser)
has a dependency on adb (an Android tool) because we end up looking for
Android devices. With this change, the Android dependency is pruned
since we determine we want to run on desktop.

Submitted with No-Presubmit to work around unrelated pylint issues.

Bug: chromium:1327303
No-Presubmit: true
Change-Id: Ice205252753ad1cd0e1ac0b4cc88e84cb6e16438
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3761392
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>