systrace: Fix manually stopping tracing in Python3
Use six.moves.input() as a stand-in for raw_input() as it is
compatible with both python2 and 3. This is conveniant as Catapult
already includes this lib in third-party, and Systrace is already
making use of it.

The general recommended approach of using "from builtins import input"
to make python3 code compatible with python2 doesn't work in this case
unless we add a new third-party dependency, so using six is simpler.

Testing:
Test with both python2 and python3
- Run: python2 systrace/bin/systrace
- Run: python3 systrace/bin/systrace
verify both commands work.

Change-Id: If26e4e4e8b2d04c71649f0206f6ce8a195b05e87
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3781216
Commit-Queue: John Reck <jreck@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: John Reck <jreck@google.com>