Devil: Add read and write permissions to on device temporary files
This CL makes the test device's shell create the temporary file and
then add the read and write permissions for all users to the file. This
will allow the android test harness to pull temporary files from the test device to the host machine.

Bug:chromium:1324649
Change-Id: I8ca4cb41cc36dc62b53ce940557c11c579fa876e
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3787093
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Rakib Hasan <rmhasan@google.com>