[pinpoint] enable python 3 fifo scheduler and refresh job tests on handlers
Fixed job related mocks. The main issue is that the mocked object in test is not the same one in sceduler. Thus, the Start() of the mocked object is actually not called. (Not sure how things work when in Python 2)

Also skipped a test related to execution engine.

Bug: chromium:1120195
Change-Id: I2fff67b4cd832382be22013c8be4251c2203f4a2
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3792282
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>