[Pinpoint] Update oauth2client version for python3 and remove usage of SignedJwtAssertionCredentials
We saw 404 error when the oauth2client v1.5.2 tried to refresh the access token. It is related to the hardcoded url in the gce workflow, which was then updated in v2.0. Meanwhile, SignedJwtAssertionCredentials was removed in v2.0 as well, thus we need to use a replacement.

The solution is referring to the following comment:
https://github.com/googleapis/oauth2client/issues/401#issuecomment-206118304

Bug: chromium:1352881
Change-Id: I3220909deb944c73bbf569314097b9ea05db20e8
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3887947
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>