[chromeperf] removed unused argument on reading cas output
The argument 'client' is not used in any of its callers. Also, even it is not None, it is not used within the function.

It is likely that we should do 'cas_client = client' in the else branch, but removing it is a safer option.

Change-Id: Ic9c257f8c12a4aab2e6189ffd5e533518b6eb3fc
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3910637
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>