Add a responsive Perfetto banner (logo + caption) to chrome://tracing.
chrome://tracing is apparently deprecated, and Perfetto is the
replacement for it. Thus, we add this banner here to let users
know that Perfetto is the recommended mechanism to use for
Chrome tracing.

For how this looks, please see the (Google-internal) screenshot:
https://screenshot.googleplex.com/Bvro6CfCvPj2Uw9

Bug: chromium:1370523
Change-Id: Id9aea51476b1bb8878d92ac842b70db252ad19cd
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3932903
Auto-Submit: Tushar Agarwal <agarwaltushar@google.com>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Andrew Shulaev <ddrone@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>