Check if the device has completed booting before doing other checks
Some of the checks in WaitUntilFullyBooted flakily timeout.
The reason this may be happening is because the device is still booting
when running these checks. This CL also changes the name of several inner functions of WaitUntilFullyBooted so they all consistently have the prefix "is_".

Bug:chromium:1286468
Change-Id: I88b7536321f71f2dfd5bbfecae60e71e1ba46b92
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3963250
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Haiyang Pan <hypan@google.com>