[chromeperf] handle OPTIONS requests in flask
The decorator for flask request handler does not handle OPTIONS request. In the existing webapp2 implementation, we do set CORS headers and return. We will do the same thing in the decorator.

OPTIONS is added automatically so we don't need to update the flask_app.route:
https://flask.palletsprojects.com/en/2.2.x/api/#flask.Flask.route

Bug: chromium:1379128
Change-Id: Idecfaebe883b1a8fb6472a4ac4b95612c488d477
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3989990
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>