Add Data Clearing Logcat Workaround
Adds the option to wait for an asynchronous intent to finish when
clearing application data. This is currently only enabled when clearing
the profile on non-rooted devices in Telemetry, but could be enabled
elsewhere if it is found to improve test stability.

Bug: chromium:1383609
Change-Id: I228e60998fc8110af96100d543cc3cea7eb29bfe
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4090563
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>