[chromeperf] Get rid the unnecessary log of "Cannot get user email as the user is not signed in"
This message shows up too frequent in upload-processing because graph_data requires extra validation for data access. It checks a few things and one of the steps is to see whether the user is signed in. Unfortunately we are relying on 'privileged request' and thus caused the 'not signed in' message to keep showing up.

Bug: chromium:1401211
Change-Id: Icb224cf58159191ab8cf073c64c62f455fff9b4c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4109360
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>