Update the flashing logic.
When flashing the userdata.img that comes with the build, the device
will only have 10GB for the internal storage size by default.
To avoid this, we instead erase the "userdata" partition and rely on
the recovery to auto format it.

Tested locally with:
* Nexus 5X
 * MMB29Q -> MMB29Q
 * MMB29Q -> N2G48C
 * N2G48C -> MMB29Q
 * N2G48C -> N2G48C
* Pixel 6
 * TP1A.220624.021 -> TP1A.220624.021

Bug: chromium:1394979, chromium:1405504
Change-Id: Icb19362d7bb1f328fb2f9eb5b0d00e0d60f55595
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4158754
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Haiyang Pan <hypan@google.com>