Refactor the test migration tool
The appengine taskqueue library works only when the handler for the task requests is running with admin mode (login:required does not work). Hence we need to move the taskqueue processing to a separate request handler that is running under admin requirements. This CL does the following.
- Moves the taskqueue handler to a separate request path that is allowed only on admin login. The landing page and initial form submission remains unchanged.
- Move the link to the test rename tool to the "Other Pages" menu since we want it to be available to non-admins too.
- Update the unauthorized page to show the bug template to use for requesting access to the tool.

Change-Id: I178e1a9681b19e15ddf905c7a3ceccd5d1cdf92c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4153279
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>