[Catapult] Add --use-persistent-shell to telemetry arg parser.
This is needed so that the option can be turned on for the
android-nougat-x86-rel bot, as it's turned on for all test suites that
the bot runs. These tests use a device, so can potentially benefit from a persistent_shell connection. android-pie-arm64-rel has been running with persistent_shell for 2 months, so shouldn't be considered a risky change anymore.

Bug: chromium:1239300
Change-Id: I2cb883f4af56fb7d8c49652c8cf6a737ffab056a
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4179050
Commit-Queue: Benjamin Joyce (Ben) <bjoyce@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Rakib Hasan <rmhasan@google.com>