Fix the unhashable error in provision_devices
Somehow the DeviceUtils instance is unhashable in python3.
As device_denylist.Read() returns a dict, python3 throws such
unhashable error when checking if a DeviceUtils instance (say "device")
exists in the dict.

Converting it to str should fix the error, and fix a potential issue
where "str(device)" is used when added to denylist, but "device"
is passed when checking the existence which will leads to no match.

Bug: chromium:1413191, chromium:1411973
Change-Id: Ieb88bc351fca3ad4710f1c7ee11aab9a65ce925f
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4226798
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Haiyang Pan <hypan@google.com>