Add remote platform TsProxyServer restart
Adds the ability to restart the TsProxyServer on remote platforms. This
is intended to be called when a test fails, as a broken SSH/adb/etc.
connection can cause a failed test and break later tests if the
TsProxyServer's forwarding is not re-established.

The GPU team has been doing this for a long time, but it's not
GPU-specific, so the potential stability improvements warrant bringing
it into Telemetry itself.

Bug: chromium:1245346
Change-Id: I0678ee34269910fb2bf59d2827bce7d725f3b40a
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4265711
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>