Pass --llvm-profile-file to Chrome execution for Lacros
In crrev/c/4358910, support for --llvm-profile-file was
introduced to allow specifying a path for profiles to
be exported to. For Lacros, we pass this variable whether
its used or not to ensure that profiles are written to /tmp
on device, which has sufficient space for profiles. It
will otherwise write to root, where the profiles will
fail to generate due to insufficient space.

Bug: chromium:1417071, chromium:1423671
Change-Id: Ieead0370ea13d06bc0d319792ac7c667e77fd3dc
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4316917
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>