[chromeperf] update auth for perf_issue_service
This CL provides a solution to add auth on the incoming requests:
 - check the request header for 'Authorization'
 - extrack the bearer token from the header
 - use the Google tokeninfo endpoint to check the token info.
 - verify the email to see  whether the request is from expect service
It will unblock the rollout of Pinpoint in prod.

However, the tokeninfo endpoint is not recommanded for being used in prod:
https://developers.google.com/identity/sign-in/web/backend-auth#using-a-google-api-client-library

We should consider using ID Token to replace it in the future.

Bug: chromium:1428666
Change-Id: Idff606f5b66a6fe9fc25e23979da74a51fc5d1c5
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4403185
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>