Add a cache for GetPcSystemType method in win_platform_backend.py
This method runs a PowerShell command to get system type information. It
is called during the ClassifyTestWithBrowser for each test case, which
makes it slow during test initialization. This CL added a cache for this
method, as the result is unlikely to change during execution.

Bug: chromium:1431374
Change-Id: I1018f95ed841c145a026a8cd94ae7757539b95ff
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4401716
Commit-Queue: Kuan Huang <kuanhuang@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>