[chromeperf] Split the current single issue tracker client into multiple that are selectable
As part of the Buganizer migration, we need different clients to talk to Monorail and Buganizer based on the project name, which tells whether it has been migrated or not. This CL renamed the existing client to be monorail client, and added a general client class.
 - The argument list of GetIssuesList is updated to make it more general. The arguments of the other calls are not changed.

Bug: chromium:1447930
Change-Id: I94bc9d5a51356cf75b8890769ddd79d4c101dcdb
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4606004
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Leina Sun <sunxiaodi@google.com>