Remove handler callback function and allow to pass job_id for testing purposes.
We remove the except stanzas. It's better to let the Workflow crash. Then when we retrieve its information through the Executions API, we will know at what exact step it failed. Since we will be using the Executions API to retrieve Sandwich results, we dont need the handler callback anymore to pass the results back to Chromeperf.

Additionally, added an optional job_id argument to the Workflow that allows user to skip the startPinpointJob step in order to save time while testing.

Bug: chromium:1419903
Change-Id: I9763245fbccd070cd79e2f3c9d4097c45278d455
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4615435
Commit-Queue: Eduardo Yap <eduardoyap@google.com>
Reviewed-by: Sean McCullough <seanmccullough@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Yuan Huang <yuanhuang@google.com>