[chromeperf] Use google.cloud.datastore to replace the google.cloud.ndb to access datastore data
The ndb libray requires a ndb.Model class to host the loaded data. If any property is missing in the class, the write operations will put null values on those missing fields. This creates dependencies between different service. E.g., the Anomaly kind is created on upload-processing during anomaly detection. When perf-issue-service tries to update the alert groups of an anomaly, it shouldn't have to know all the other properties of the anomaly, such as median_before_anomaly.

Now I change to use the google.cloud.datastore model which represents the entities in json wich does not require knowledges of the existing fields.

Bug: chromium:1468326
Change-Id: I8962c19d05af168ca64c48fa03b642f07530aefe
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4724123
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>