[devil] Dump netstat when out of ports
Updates devil's forwarder to automatically dump the output of netstat if
a failure occurs and the logs indicate that it could be due to the
device running out of ports. Currently, a user will be able to tell when
this happens due to the logs we already surface, but won't be able to do
much with the information since they won't be able to check what is
taking up all the available ports.

Bug: dawn:2024
Change-Id: I2dd3ef981e361f2b247f2412542162ffff8dff1c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4827114
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>