devil: Add a new method GetShellWritablePath
When the active user on device is a secondary user, the paths like
"/sdcard" and "/data/data" points to different paths case by case.

For example:
 * when in "am instrument", they points to the secondary user's
   sdcard and app storage.
 * when in "adb shell", they always points to the system user 0's
   sdcard and app storage.

This method converts a device path of a target user to one that is
accessible by the adb shell commands.

Also update DeviceTempFile to take DeviceUtils as an optional param,
as using adb to create/delete is not sufficient for a secondary user
due to the above path change.

Bug: b:293175593
Change-Id: I7780d665e923ccf55ce506520732379aa83cf663
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4827105
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Haiyang Pan <hypan@google.com>