[chromeperf] fix the missing labels on delay triage workflow
Fix the issue that the required labels are removed, such as Pri-X and Type-X, by addding a parameter when generating the components, cc and labels.
Also allow subscriptions to opt-out from the delay triage. Currently sandwich subs will still be triaged as the label 'sandwich-sheriff' is needed.

Bug: chromium:1471690
Change-Id: I35747833dba7ddec82f1ca5047625d00f27c4a6a
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4863928
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>