Fix generic-set-span style
the "a" style element doesn't appear to get applied to the links
as intended.

Also this style labeled "a" also has an implement dependency on the
internal implementation of createLink() which creates an element
named a.

Fix both of thse issues by setting the style at the time of link
creation.

This will have the impact of making sure the values are not "squashed"
together like "0123" and instead appear on separate lines as originally
intended like:

```
0
1
2
3
```

Change-Id: Ice104cb45f23ce3c318cf5edf489f5437c852a63
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4977696
Reviewed-by: Deep Roy <dproy@chromium.org>
Commit-Queue: Deep Roy <dproy@chromium.org>