Avoid Android dialog during benchmarking
When the POST_NOTIFICATION android permission is not granted for Chrome,
a notification pops up that obscures much of the screen and darkens the
rest of the screen.

There is no simple way (e.g. command-line flag) to avoid the popup, so
the simplest alternative is to grant the permission when setting up the
environment. This allows performance tests using run_benchmark to
proceed properly.

Bug: chromium:1498208
Change-Id: I1f689bb97c849f0751078e5576378cf9e0f57a12
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/4995555
Auto-Submit: Peter Wen <wnwen@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Peter Wen <wnwen@chromium.org>