[chromeperf] reduce task processing rate for group update tasks
Currently we process ~2000 groups per 10 minutes. The rate is 20/s which tries to execute all tasks in 100 seconds, which caused a burst of processing at the begining of the 10 minute cycle.

This CL is:
- reducing the rate to 10/s to reduce processing rate.
- max_concurrent_request to 100 to reduce database contention.
- the retry limit to 1 and back off by 120 seconds to retry after peak.

Also, the interval of the group processing has been reduced from every 10 mins to every 15 mins.

Change-Id: Ic4303586b1da4fefb66225e368b53bef7a6458a7
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5046340
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>