[chromeperf] Update the buganizer client to support Fuchsia
We have another CL added a new test subscription:
https://chrome-internal-review.googlesource.com/c/infra/infra_internal/+/6718102

With that, the same anomaly will be mapped to two subscriptions, and thus two issue will be filed: one in buganizer and one in monorail.

This CL updates the client so that it can take the inputs of the new subscription.

Bug: chromium:1447930
Change-Id: I496913a073cf05ec2feaff7cca182356962f000c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5075114
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>