[chromeperf] buganizer client return the issue id as int
The new issue response from Buganizer has the issue id as string instead of integer which is from Monorail. This causes a failure when initializing the BugInfo object as it expect the id as an datastore IntegerProperty.

This causes a problem that after an issue is created, the BugInfo cannot be created and thus the group status is not updated to be 'triaged'. As a result, multiple issues are being created on the same anomaly group.

This CL convert the value before sending the response.
(A more comprehensive fix would be convering on the alert group workflow before creating the BugInfo object. Though, rolling out the perf-issue-service is more light-weight and has less impact. Another CL will be done on the default service)

Bug: chromium:1447930
Change-Id: Ia46970c5ddf606396f12b76f1523e1d0e70f7da8
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5080419
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>