[Pinpoint] disable culprit verification for non-chromium culprits
When a culprit is from a non-chromium repo, culprit verification will
start on the nearest chromium commit prior, creating an A/A job.
The ideal solution is to run the culprit-1 vs culprit but if the
culprit-1 and culprit are part of different repos, then Pinpoint
cannot support the job request. Until Pinpoint can properly support
this case, we disable culprit verification for non-chromium commits.
This case occurs about 30% of the time.

Bug: chromium:1507128
Change-Id: I8dc787e13e4b47b7c07ae84a704d86431062143b
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5085148
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Leina Sun <sunxiaodi@google.com>