[chromeperf] If bug id is big enough, use buganizer client regardless of the project name
Use issue id to decide whether to use buganizer client. This workaround is needed during the validation phase that Pinpoint will take a buganizer issue ID as input. Because 'chromium' project is not migrated yet, we need to rely on the ID to choose the Buganizer client.

Bug: chromium:1447930
Change-Id: I2d5564c24e331a3f7fdc14eab36ab9c3ad9bc657
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5113002
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>