devil: only check expected devices for recovery and status.
The scripts used to check the union of expected devices and devices
returned by lsusb.
However lsusb started return all the devices from the host, instead of
the assigned one in the docker container. This leads to the scripts to
report missing for the rest devices, as they are not reachable from the docker container. See crbug.com/1469007#c5 for details.

Update the logic from "union" to "intersection" so that the scripts only
checks the device assigned in the docker container as it is the one we
really care about.

Bug: chromium:1395189
Change-Id: Iaf97f120600c97dfcbcd46a9444fcf2cd7bfaa5a
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5124441
Commit-Queue: Haiyang Pan <hypan@google.com>
Reviewed-by: Egor Pasko <pasko@chromium.org>