[pinpoint] Support DEPS parsing endpoint
The ability to parse DEPS files through gclient_eval (depot_tools)
is not reproducible in Go with the ast library.

During bisection, when two commits are adjacent to each other,
a DEPS roll is assumed. A DEPS roll mostly rolls one repository.
Bisection requires DEPS parsing to determine the repository and
revision that's changed between the two adjacent commits to determine
the next potential candidate to bisect.

The response is formatted from the frozenset() into JSON dict format,
where key is repository_url and value is the revision.

DEPS is currently in a hybrid state with `git submodules`, and there
are eventual plans to remove .git based repositories from DEPS. This
endpoint would likely be deprecated as a new solution is developed
using the submodules.

Change-Id: I40320c14b0a5c92f40626a1723f0e6d413a7774e
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5146595
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>