[skia-bridge] Add DEPS parsing endpoint
We've exposed a DEPS parsing endpoint in Pinpoint, but with
Pinpoint moving towards deprecation it made more sense to have this
supported in skia-bridge.

This is expected to be a stop-gap until a permanent solution for DEPS
is provided.

See crrev/c/5146595 for original change and motive in Pinpoint.

TEST = Running app locally and calling new endpoint.

Change-Id: I52da375c74fee7de706ef94cf9b33fe154873896
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5160992
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>