[Chromeperf] Update handling for monorail labels in buganizer client
Buganizer only provides hotlist mapping for some of the Monorail labels. During migration, the labels with no mappings will be copied to a custom field in the Buganizer issue. The value of the custom field is 'Monorail Labels' for project Fuchsia, and 'Chromium Labels' for the other projects.

Also, the label update in Buganizer is different from Monorail that the new value in the update request will simply override the current value. Thus, we need to load the original values and add/remove from there.

Bug: chromium:1447930
Change-Id: Ib5b890ab6ad7eae83c60a989eaa024eb0fb71a2c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5203608
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>