Support AddMeasurement on Story class
In some cases we need to collect ad-hoc measurements on Story classes
that do not derive from PressStory. This change allows us to do so.

The code was copied from PressStory. In a followup we should be able
to remove that copy.

Change-Id: I1cbb42fec0d9c2e6e20a41be364fa215c587f591
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5208133
Auto-Submit: Peter Collingbourne <pcc@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>