[chromeperf] remove assignee if the issue marked as duplicate
In buganizer, setting duplicate uses a different call before updating the issue. If the 'assignee' field is not empty, the issue status will be set to 'assigned', which overwrites the 'duplicate' status just set.

This CL removes the 'assignee' field if the issue is being set duplicate.

Bug: chromium:324005871
Change-Id: I38420d2cf3b0e5b4fe19ea67e0d32ed5c2a1080c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5271259
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>