[chromeperf] fix the label list passed as a string
Labels are expected to be paased as a list. If it is passed as string, it will be considered as a list of chars. This CL fixes a few locations that a single label is sent as string.

Bug: chromium:324276740
Change-Id: I503726b13384d3d9b721f79cbc79caf1b43d2bf3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5276714
Auto-Submit: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>