[chromeperf] fix the issue that comment id can be null
In some cases, the 'version' from the issue update can be null. Currently we are using 'version' value as 'id'. The null value can cause error when we sort by the 'id'.

This CL change the logic to use the issue update's index as the 'id'.

bug: chromium:324359771
Change-Id: Ie36c4bdd98cc203f84b7fc7fff658224b7ad9f26
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5276715
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>