Fix race condition when creating new tab
When telemetry is creating new tab it sequentially uses two devtools
endpoints:
1. websocket Target.createTarget - to create tab and get target_id
2. REST '' - to get list of available contexts

Telemetry is expecting, that at the moment of making request number 2
target_id from response of request number 1 exists in a list of
available contexts. But actual response of request number 2 contains
target_id from time to time.

This happens because of race condition - at the moment of
sending request number 2 target_id is still missing from the list of
contexts. Waiting for some time between request number 1 and request
number 2 solves this problem.

On my computer timeout of 0.1 s or 0.01 s is enough. We use 1 s as a
timeout to make sure that this timeout is enough for majority of
computers, even old ones.

Change-Id: I579cd499e8528c165515fb6132a67245c8c4d1ba
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5454397
Commit-Queue: Dmitry Shcherbinin <cher-di@yandex-team.ru>
Reviewed-by: John Chen <johnchen@chromium.org>