[pinpoint] filter small regressions from bisection
Bisections triggered by anomaly detection are typically launched with
a comparison magnitude that measures the expected regression size.
Bisection is capable of finding culprits for regressions that are
much smaller in a given commit range but this leads to wasted resources
and extra noise for developers.

This CL filters out regressions that are below a certain threshold
relative to the initial comparison magnitude.

Bug: chromium:340884347
Change-Id: I1411fcb50feaba5558c3c77716be547326caa17a
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5560316
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>