[chromeperf] Pinpoint raise error when it fails to parse DEPS file
The third_party/depot_tools is out of sync and started to fail when parsing the DEPS files. This error was not raised and has been considered as 'no DEPS', which is not correct. It caused the bisection to stop drill down into the sub repositories while it should.

This CL raise error for the DEPS parsing error. While we will have another CL to fix the parsing.

Bug: chromium:343229141
Change-Id: I2dbfe92b1dda3a3904d280c5eed3e69a1575aea8
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5617491
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>