[fuchsia] use build/fuchsia in fuchsia_interface
The implementation needs to be shared across multiple fuchsia
test facilities, so the change is a prerequisite of moving the
code to build/fuchsia.

This change is verified by https://crrev.com/c/5747116.

I do believe
https://crsrc.org/c/tools/perf/process_perf_results.pydeps
needs to be updated. But the dynamic importing would give us
some flexibility to do it after the rolling.

Bug: chromium:40935291
Change-Id: I8f28ec7f2d278f762137578ea05aeed2f95a2271
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5747420
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Zijie He <zijiehe@google.com>