Fix _RunAdbCmd timeout logic
Fixes the timeout logic in _RunAdbCmd to properly handle the case when
a timeout is specified.

Currently, the logic is:

if timeout:
  calculate the timeout to use (which doesn't involve |timeout|)
else:
  use |timeout|

This appears to be the opposite of what we want, and is suspected of
contributing to a failure to retry certain actions in ANGLE tests.

Bug: chromium:339457761, b:344849382
Change-Id: Ic65ac0fedd0fd187d958b48a03f6e26bd49fb649
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5807295
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>