[chromeperf] add file_bug_skia to use oauth and add logs
The current bisect request from Skia arrives at Pinpoint's api/new endpoint, which satisfies the IsTryJobUser() check, which should also satisfy the IsValidSheriffUser().

However, the request from Skia to Chromeperf's file_bug_skia/ failed the isValidSheriffUser() check. The difference could be in the GetEmail() call, which validates the user in different ways based on the request uri.

Adding the file_bug_skia/ to the OAUTH_ENDPOINTS allign it with api/. Ideally we should put the file_bug_skia/ under api/ as well. We will do this once we verify the cause of the failure.

Bug: b:365008653
Change-Id: I7a0fd5baa7e038316148a48f266a2915a6846ebd
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/5874068
Reviewed-by: Eduardo Yap <eduardoyap@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>